backend/requirements.txt
flask
flask-cors
psycopg2
python-dotenv
supabase
ultralytics
